Refactor: Move `AudioService` and `AudioTrack` to dedicated packages
- Moved `AudioService.kt` to `service` package.
- Moved `AudioTrack.kt` to `data` package.
- Moved `AudioAdapter.kt` to `ui` package.
- Updated imports in `MainActivity.kt`.
- Updated service name in `AndroidManifest.xml` to reflect new package.
- Updated Gradle version.
