+++ app/src/main/java/com/grupo11/equalizador/data/AudioTrack.kt
@@ -1,0 +1,6 @@
+package com.grupo11.equalizador.data
+
+data class AudioTrack(
+    val title: String,   // nome para exibição
+    val resId: Int       // ID do recurso raw (R.raw.xxx)
+)
